Fixed multiple groupby execution in single pipeline
