+++ src/core/etl/tests/Flow/ETL/Tests/Unit/DataFrameTest.php
@@ -6,14 +6,9 @@ namespace Flow\ETL\Tests\Unit;
 
 use function Flow\ETL\DSL\average;
 use function Flow\ETL\DSL\lit;
-use function Flow\ETL\DSL\max;
-use function Flow\ETL\DSL\rank;
 use function Flow\ETL\DSL\ref;
 use function Flow\ETL\DSL\refs;
-use function Flow\ETL\DSL\sum;
-use function Flow\ETL\DSL\window;
 use Flow\ETL\DataFrame;
-use Flow\ETL\DataFrameFactory;
 use Flow\ETL\DSL\Entry;
 use Flow\ETL\DSL\From;
 use Flow\ETL\DSL\Partitions;
@@ -24,32 +19,17 @@ use Flow\ETL\Exception\InvalidArgumentException;
 use Flow\ETL\Extractor;
 use Flow\ETL\Flow;
 use Flow\ETL\FlowContext;
-use Flow\ETL\Join\Expression;
 use Flow\ETL\Loader;
-use Flow\ETL\Memory\ArrayMemory;
-use Flow\ETL\PHP\Type\Logical\List\ListElement;
-use Flow\ETL\PHP\Type\Logical\ListType;
-use Flow\ETL\PHP\Type\Logical\Map\MapKey;
-use Flow\ETL\PHP\Type\Logical\Map\MapValue;
-use Flow\ETL\PHP\Type\Logical\MapType;
-use Flow\ETL\PHP\Type\Logical\Structure\StructureElement;
-use Flow\ETL\PHP\Type\Logical\StructureType;
-use Flow\ETL\PHP\Type\Native\ScalarType;
 use Flow\ETL\Row;
 use Flow\ETL\Row\Entry\ArrayEntry;
 use Flow\ETL\Row\Entry\BooleanEntry;
 use Flow\ETL\Row\Entry\DateTimeEntry;
-use Flow\ETL\Row\Entry\FloatEntry;
 use Flow\ETL\Row\Entry\IntegerEntry;
-use Flow\ETL\Row\Entry\ListEntry;
-use Flow\ETL\Row\Entry\MapEntry;
 use Flow\ETL\Row\Entry\NullEntry;
 use Flow\ETL\Row\Entry\StringEntry;
-use Flow\ETL\Row\Entry\StructureEntry;
 use Flow\ETL\Row\Schema;
 use Flow\ETL\Rows;
 use Flow\ETL\Tests\Double\AddStampToStringEntryTransformer;
-use Flow\ETL\Tests\Fixtures\Enum\BackedStringEnum;
 use Flow\ETL\Transformation;
 use Flow\ETL\Transformer;
 use Flow\ETL\Transformer\StyleConverter\StringStyles;
@@ -119,47 +99,6 @@ final class DataFrameTest extends TestCase
             ->count();
 
         $this->assertSame(5, $count);
-    }
-
-    public function test_cross_join() : void
-    {
-        $loader = $this->createMock(Loader::class);
-        $loader->expects($this->exactly(2))
-            ->method('load');
-
-        $rows = (new Flow())->process(
-            new Rows(
-                Row::create(Entry::integer('id', 1), Entry::string('country', 'PL')),
-                Row::create(Entry::integer('id', 2), Entry::string('country', 'PL')),
-                Row::create(Entry::integer('id', 3), Entry::string('country', 'PL')),
-                Row::create(Entry::integer('id', 4), Entry::string('country', 'PL')),
-            )
-        )
-            ->batchSize(2)
-            ->crossJoin(
-                (new Flow())->process(
-                    new Rows(
-                        Row::create(Entry::integer('num', 1), Entry::boolean('active', true)),
-                        Row::create(Entry::integer('num', 2), Entry::boolean('active', false)),
-                    )
-                ),
-            )
-            ->write($loader)
-            ->fetch();
-
-        $this->assertEquals(
-            new Rows(
-                Row::create(Entry::integer('id', 1), Entry::string('country', 'PL'), Entry::integer('num', 1), Entry::boolean('active', true)),
-                Row::create(Entry::integer('id', 1), Entry::string('country', 'PL'), Entry::integer('num', 2), Entry::boolean('active', false)),
-                Row::create(Entry::integer('id', 2), Entry::string('country', 'PL'), Entry::integer('num', 1), Entry::boolean('active', true)),
-                Row::create(Entry::integer('id', 2), Entry::string('country', 'PL'), Entry::integer('num', 2), Entry::boolean('active', false)),
-                Row::create(Entry::integer('id', 3), Entry::string('country', 'PL'), Entry::integer('num', 1), Entry::boolean('active', true)),
-                Row::create(Entry::integer('id', 3), Entry::string('country', 'PL'), Entry::integer('num', 2), Entry::boolean('active', false)),
-                Row::create(Entry::integer('id', 4), Entry::string('country', 'PL'), Entry::integer('num', 1), Entry::boolean('active', true)),
-                Row::create(Entry::integer('id', 4), Entry::string('country', 'PL'), Entry::integer('num', 2), Entry::boolean('active', false)),
-            ),
-            $rows
-        );
     }
 
     public function test_drop() : void
@@ -252,284 +191,7 @@ final class DataFrameTest extends TestCase
         );
     }
 
-    public function test_etl() : void
-    {
-        $extractor = new class implements Extractor {
-            /**
-             * @param FlowContext $context
-             *
-             * @return \Generator<int, Rows, mixed, void>
-             */
-            public function extract(FlowContext $context) : \Generator
-            {
-                yield new Rows(
-                    Row::create(
-                        new IntegerEntry('id', 101),
-                        new BooleanEntry('deleted', false),
-                        new DateTimeEntry('expiration-date', new \DateTimeImmutable('2020-08-24')),
-                        new NullEntry('phase')
-                    )
-                );
-
-                yield new Rows(
-                    Row::create(
-                        new IntegerEntry('id', 102),
-                        new BooleanEntry('deleted', true),
-                        new DateTimeEntry('expiration-date', new \DateTimeImmutable('2020-08-25')),
-                        new NullEntry('phase')
-                    )
-                );
-            }
-        };
-
-        $addStampStringEntry = new class implements Transformer {
-            public function transform(Rows $rows, FlowContext $context) : Rows
-            {
-                return $rows->map(
-                    fn (Row $row) : Row => $row->set(new StringEntry('stamp', 'zero'))
-                );
-            }
-
-            public function __serialize() : array
-            {
-                return [];
-            }
-
-            public function __unserialize(array $data) : void
-            {
-            }
-        };
-
-        $loader = new class implements Loader {
-            public array $result = [];
-
-            public function load(Rows $rows, FlowContext $context) : void
-            {
-                $this->result = \array_merge($this->result, $rows->toArray());
-            }
-
-            public function __serialize() : array
-            {
-                return [];
-            }
-
-            public function __unserialize(array $data) : void
-            {
-            }
-        };
-
-        (new Flow())->read($extractor)
-            ->onError(new IgnoreError())
-            ->rows($addStampStringEntry)
-            ->rows(new class implements Transformer {
-                public function transform(Rows $rows, FlowContext $context) : Rows
-                {
-                    throw new \RuntimeException('Unexpected exception');
-                }
-
-                public function __serialize() : array
-                {
-                    return [];
-                }
-
-                public function __unserialize(array $data) : void
-                {
-                }
-            })
-            ->rows(AddStampToStringEntryTransformer::divideBySemicolon('stamp', 'one'))
-            ->rows(AddStampToStringEntryTransformer::divideBySemicolon('stamp', 'two'))
-            ->rows(AddStampToStringEntryTransformer::divideBySemicolon('stamp', 'three'))
-            ->write($loader)
-            ->run();
-
-        $this->assertEquals(
-            [
-                [
-                    'id' => 101,
-                    'stamp' => 'zero:one:two:three',
-                    'deleted' => false,
-                    'expiration-date' => new \DateTimeImmutable('2020-08-24'),
-                    'phase' => null,
-                ],
-                [
-                    'id' => 102,
-                    'stamp' => 'zero:one:two:three',
-                    'deleted' => true,
-                    'expiration-date' => new \DateTimeImmutable('2020-08-25'),
-                    'phase' => null,
-                ],
-            ],
-            $loader->result,
-        );
-    }
-
-    public function test_etl_display() : void
-    {
-        $etl = (new Flow())->extract(
-            new class implements Extractor {
-                /**
-                 * @return \Generator<int, Rows, mixed, void>
-                 */
-                public function extract(FlowContext $context) : \Generator
-                {
-                    for ($i = 0; $i < 20; $i++) {
-                        yield new Rows(
-                            Row::create(
-                                new IntegerEntry('id', 1234),
-                                new FloatEntry('price', 123.45),
-                                new IntegerEntry('100', 100),
-                                new BooleanEntry('deleted', false),
-                                new DateTimeEntry('created-at', new \DateTimeImmutable('2020-07-13 15:00')),
-                                new NullEntry('phase'),
-                                new ArrayEntry(
-                                    'array',
-                                    [
-                                        ['id' => 1, 'status' => 'NEW'],
-                                        ['id' => 2, 'status' => 'PENDING'],
-                                    ]
-                                ),
-                                new ListEntry(
-                                    'list',
-                                    [1, 2, 3],
-                                    new ListType(ListElement::integer())
-                                ),
-                                new MapEntry(
-                                    'map',
-                                    ['NEW', 'PENDING'],
-                                    new MapType(MapKey::integer(), MapValue::string())
-                                ),
-                                new StructureEntry(
-                                    'items',
-                                    ['item-id' => '1', 'name' => 'one'],
-                                    new StructureType(
-                                        new StructureElement('item-id', ScalarType::string()),
-                                        new StructureElement('name', ScalarType::string()),
-                                    )
-                                ),
-                                new Row\Entry\ObjectEntry('object', new \ArrayIterator([1, 2, 3])),
-                                new Row\Entry\EnumEntry('enum', BackedStringEnum::three),
-                                new Row\Entry\XMLEntry('xml', '<xml><node id="123">test<foo>bar</foo></node></xml>'),
-                            ),
-                        );
-                    }
-                }
-            }
-        );
-
-        $this->assertSame(
-            <<<'ASCIITABLE'
-+------+--------+-----+---------+----------------------+-------+----------------------+---------+-------------------+----------------------+----------------------+-------+----------------------+
-|   id |  price | 100 | deleted |           created-at | phase |                array |    list |               map |                items |               object |  enum |                  xml |
-+------+--------+-----+---------+----------------------+-------+----------------------+---------+-------------------+----------------------+----------------------+-------+----------------------+
-| 1234 | 123.45 | 100 |   false | 2020-07-13T15:00:00+ |  null | [{"id":1,"status":"N | [1,2,3] | ["NEW","PENDING"] | {"item-id":"1","name | ArrayIterator Object | three | <?xml version="1.0"? |
-| 1234 | 123.45 | 100 |   false | 2020-07-13T15:00:00+ |  null | [{"id":1,"status":"N | [1,2,3] | ["NEW","PENDING"] | {"item-id":"1","name | ArrayIterator Object | three | <?xml version="1.0"? |
-| 1234 | 123.45 | 100 |   false | 2020-07-13T15:00:00+ |  null | [{"id":1,"status":"N | [1,2,3] | ["NEW","PENDING"] | {"item-id":"1","name | ArrayIterator Object | three | <?xml version="1.0"? |
-| 1234 | 123.45 | 100 |   false | 2020-07-13T15:00:00+ |  null | [{"id":1,"status":"N | [1,2,3] | ["NEW","PENDING"] | {"item-id":"1","name | ArrayIterator Object | three | <?xml version="1.0"? |
-| 1234 | 123.45 | 100 |   false | 2020-07-13T15:00:00+ |  null | [{"id":1,"status":"N | [1,2,3] | ["NEW","PENDING"] | {"item-id":"1","name | ArrayIterator Object | three | <?xml version="1.0"? |
-+------+--------+-----+---------+----------------------+-------+----------------------+---------+-------------------+----------------------+----------------------+-------+----------------------+
-5 rows
-
-ASCIITABLE,
-            $etl->display(5)
-        );
-
-        $this->assertSame(
-            <<<'ASCIITABLE'
-+------+--------+-----+---------+---------------------------+-------+-------------------------------------------------------+---------+-------------------+------------------------------+------------------------------------------------------------------------------------------------+-------+--------------------------------------------------------------------------+
-|   id |  price | 100 | deleted |                created-at | phase |                                                 array |    list |               map |                        items |                                                                                         object |  enum |                                                                      xml |
-+------+--------+-----+---------+---------------------------+-------+-------------------------------------------------------+---------+-------------------+------------------------------+------------------------------------------------------------------------------------------------+-------+--------------------------------------------------------------------------+
-| 1234 | 123.45 | 100 |   false | 2020-07-13T15:00:00+00:00 |  null | [{"id":1,"status":"NEW"},{"id":2,"status":"PENDING"}] | [1,2,3] | ["NEW","PENDING"] | {"item-id":"1","name":"one"} | ArrayIterator Object( [storage:ArrayIterator:private] => Array ( [0] => 1 [1] => 2 [2] => 3 )) | three | <?xml version="1.0"?><xml><node id="123">test<foo>bar</foo></node></xml> |
-| 1234 | 123.45 | 100 |   false | 2020-07-13T15:00:00+00:00 |  null | [{"id":1,"status":"NEW"},{"id":2,"status":"PENDING"}] | [1,2,3] | ["NEW","PENDING"] | {"item-id":"1","name":"one"} | ArrayIterator Object( [storage:ArrayIterator:private] => Array ( [0] => 1 [1] => 2 [2] => 3 )) | three | <?xml version="1.0"?><xml><node id="123">test<foo>bar</foo></node></xml> |
-| 1234 | 123.45 | 100 |   false | 2020-07-13T15:00:00+00:00 |  null | [{"id":1,"status":"NEW"},{"id":2,"status":"PENDING"}] | [1,2,3] | ["NEW","PENDING"] | {"item-id":"1","name":"one"} | ArrayIterator Object( [storage:ArrayIterator:private] => Array ( [0] => 1 [1] => 2 [2] => 3 )) | three | <?xml version="1.0"?><xml><node id="123">test<foo>bar</foo></node></xml> |
-| 1234 | 123.45 | 100 |   false | 2020-07-13T15:00:00+00:00 |  null | [{"id":1,"status":"NEW"},{"id":2,"status":"PENDING"}] | [1,2,3] | ["NEW","PENDING"] | {"item-id":"1","name":"one"} | ArrayIterator Object( [storage:ArrayIterator:private] => Array ( [0] => 1 [1] => 2 [2] => 3 )) | three | <?xml version="1.0"?><xml><node id="123">test<foo>bar</foo></node></xml> |
-| 1234 | 123.45 | 100 |   false | 2020-07-13T15:00:00+00:00 |  null | [{"id":1,"status":"NEW"},{"id":2,"status":"PENDING"}] | [1,2,3] | ["NEW","PENDING"] | {"item-id":"1","name":"one"} | ArrayIterator Object( [storage:ArrayIterator:private] => Array ( [0] => 1 [1] => 2 [2] => 3 )) | three | <?xml version="1.0"?><xml><node id="123">test<foo>bar</foo></node></xml> |
-| 1234 | 123.45 | 100 |   false | 2020-07-13T15:00:00+00:00 |  null | [{"id":1,"status":"NEW"},{"id":2,"status":"PENDING"}] | [1,2,3] | ["NEW","PENDING"] | {"item-id":"1","name":"one"} | ArrayIterator Object( [storage:ArrayIterator:private] => Array ( [0] => 1 [1] => 2 [2] => 3 )) | three | <?xml version="1.0"?><xml><node id="123">test<foo>bar</foo></node></xml> |
-+------+--------+-----+---------+---------------------------+-------+-------------------------------------------------------+---------+-------------------+------------------------------+------------------------------------------------------------------------------------------------+-------+--------------------------------------------------------------------------+
-6 rows
-
-ASCIITABLE,
-            $etl->display(6, 0)
-        );
-    }
-
-    public function test_etl_display_with_very_long_entry_name() : void
-    {
-        $etl = (new Flow())
-            ->read(
-                From::array([
-                [
-                    'this is very long entry name that should be longer than items' => [
-                        ['id' => 1, 'status' => 'NEW'],
-                        ['id' => 2, 'status' => 'PENDING'],
-                    ],
-                ],
-                [
-                    'this is very long entry name that should be longer than items' => [
-                        ['id' => 1, 'status' => 'NEW'],
-                        ['id' => 2, 'status' => 'PENDING'],
-                    ],
-                ],
-                [
-                    'this is very long entry name that should be longer than items' => [
-                        ['id' => 1, 'status' => 'NEW'],
-                        ['id' => 2, 'status' => 'PENDING'],
-                    ],
-                ],
-                [
-                    'this is very long entry name that should be longer than items' => [
-                        ['id' => 1, 'status' => 'NEW'],
-                        ['id' => 2, 'status' => 'PENDING'],
-                    ],
-                ],
-                [
-                    'this is very long entry name that should be longer than items' => [
-                        ['id' => 1, 'status' => 'NEW'],
-                        ['id' => 2, 'status' => 'PENDING'],
-                    ],
-                ],
-                [
-                    'this is very long entry name that should be longer than items' => [
-                        ['id' => 1, 'status' => 'NEW'],
-                        ['id' => 2, 'status' => 'PENDING'],
-                    ],
-                ],
-            ])
-            );
-
-        $this->assertStringContainsString(
-            <<<'ASCIITABLE'
-+----------------------+
-| this is very long en |
-+----------------------+
-| [{"id":1,"status":"N |
-| [{"id":1,"status":"N |
-| [{"id":1,"status":"N |
-| [{"id":1,"status":"N |
-| [{"id":1,"status":"N |
-+----------------------+
-5 rows
-ASCIITABLE,
-            $etl->display(5)
-        );
-
-        $this->assertStringContainsString(
-            <<<'ASCIITABLE'
-+---------------------------------------------------------------+
-| this is very long entry name that should be longer than items |
-+---------------------------------------------------------------+
-|         [{"id":1,"status":"NEW"},{"id":2,"status":"PENDING"}] |
-|         [{"id":1,"status":"NEW"},{"id":2,"status":"PENDING"}] |
-|         [{"id":1,"status":"NEW"},{"id":2,"status":"PENDING"}] |
-|         [{"id":1,"status":"NEW"},{"id":2,"status":"PENDING"}] |
-|         [{"id":1,"status":"NEW"},{"id":2,"status":"PENDING"}] |
-+---------------------------------------------------------------+
-5 rows
-ASCIITABLE,
-            $etl->display(5, 0)
-        );
-    }
-
-    public function test_etl_exceeding_the_limit_in_one_rows_set() : void
+    public function test_exceeding_the_limit_in_one_rows_set() : void
     {
         $rows = (new Flow())
             ->read(
@@ -544,7 +206,7 @@ ASCIITABLE,
         $this->assertCount(9, $rows);
     }
 
-    public function test_etl_fetch_with_limit() : void
+    public function test_fetch_with_limit() : void
     {
         $rows = (new Flow())
             ->read(From::array([
@@ -564,7 +226,16 @@ ASCIITABLE,
         $this->assertCount(5, $rows);
     }
 
-    public function test_etl_fetch_without_limit() : void
+    public function test_fetch_with_limit_below_0() : void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage("Limit can't be lower or equal zero, given: -1");
+
+        (new Flow())->process(new Rows())
+            ->fetch(-1);
+    }
+
+    public function test_fetch_without_limit() : void
     {
         $rows = (new Flow())->extract(
             new class implements Extractor {
@@ -588,7 +259,7 @@ ASCIITABLE,
         $this->assertCount(20, $rows);
     }
 
-    public function test_etl_filter() : void
+    public function test_filter() : void
     {
         $rows = (new Flow())->extract(
             new class implements Extractor {
@@ -615,296 +286,6 @@ ASCIITABLE,
             [['id' => 2], ['id' => 4], ['id' => 6], ['id' => 8], ['id' => 10]],
             $rows->toArray()
         );
-    }
-
-    public function test_etl_limit() : void
-    {
-        $rows = (new Flow())->extract(
-            new class implements Extractor {
-                /**
-                 * @param FlowContext $context
-                 *
-                 * @return \Generator<int, Rows, mixed, void>
-                 */
-                public function extract(FlowContext $context) : \Generator
-                {
-                    for ($i = 0; $i < 1000; $i++) {
-                        yield new Rows(
-                            Row::create(new IntegerEntry('id', $i + 1)),
-                            Row::create(new IntegerEntry('id', $i + 2)),
-                        );
-                    }
-                }
-            }
-        )
-        ->limit(10)
-        ->fetch();
-
-        $this->assertCount(10, $rows);
-    }
-
-    public function test_etl_limit_with_batch_size() : void
-    {
-        $rows = (new Flow())->extract(
-            new class implements Extractor {
-                /**
-                 * @param FlowContext $context
-                 *
-                 * @return \Generator<int, Rows, mixed, void>
-                 */
-                public function extract(FlowContext $context) : \Generator
-                {
-                    for ($i = 0; $i < 1000; $i++) {
-                        yield new Rows(
-                            Row::create(new IntegerEntry('id', $i + 1)),
-                            Row::create(new IntegerEntry('id', $i + 2)),
-                        );
-                    }
-                }
-            }
-        )
-            ->batchSize(50)
-            ->limit(10)
-            ->fetch();
-
-        $this->assertCount(10, $rows);
-    }
-
-    public function test_etl_limit_with_collecting() : void
-    {
-        $rows = (new Flow())->extract(
-            new class implements Extractor {
-                /**
-                 * @param FlowContext $context
-                 *
-                 * @return \Generator<int, Rows, mixed, void>
-                 */
-                public function extract(FlowContext $context) : \Generator
-                {
-                    for ($i = 0; $i < 1000; $i++) {
-                        yield new Rows(
-                            Row::create(new IntegerEntry('id', $i + 1)),
-                            Row::create(new IntegerEntry('id', $i + 2)),
-                        );
-                    }
-                }
-            }
-        )
-            ->limit(10)
-            ->collect()
-            ->fetch();
-
-        $this->assertCount(10, $rows);
-    }
-
-    public function test_etl_map() : void
-    {
-        $rows = (new Flow())->extract(
-            new class implements Extractor {
-                /**
-                 * @param FlowContext $context
-                 *
-                 * @return \Generator<int, Rows, mixed, void>
-                 */
-                public function extract(FlowContext $context) : \Generator
-                {
-                    for ($i = 1; $i <= 10; $i++) {
-                        yield new Rows(
-                            Row::create(new IntegerEntry('id', $i)),
-                        );
-                    }
-                }
-            }
-        )
-            ->map(fn (Row $row) => $row->add(new BooleanEntry('odd', $row->valueOf('id') % 2 === 0)))
-            ->fetch();
-
-        $this->assertCount(10, $rows);
-        $this->assertSame(
-            [
-                ['id' => 1, 'odd' => false],
-                ['id' => 2, 'odd' => true],
-                ['id' => 3, 'odd' => false],
-                ['id' => 4, 'odd' => true],
-                ['id' => 5, 'odd' => false],
-                ['id' => 6, 'odd' => true],
-                ['id' => 7, 'odd' => false],
-                ['id' => 8, 'odd' => true],
-                ['id' => 9, 'odd' => false],
-                ['id' => 10, 'odd' => true],
-            ],
-            $rows->toArray()
-        );
-    }
-
-    public function test_etl_process_constructor() : void
-    {
-        $collectedRows = (new Flow())->process(
-            $rows = new Rows(
-                Row::create(
-                    new IntegerEntry('id', 101),
-                    new BooleanEntry('deleted', false),
-                    new DateTimeEntry('expiration-date', new \DateTimeImmutable('2020-08-24')),
-                    new NullEntry('phase')
-                )
-            )
-        )
-            ->fetch();
-
-        $this->assertEquals($rows, $collectedRows);
-    }
-
-    public function test_etl_with_batch_size() : void
-    {
-        (new Flow())->extract(
-            new class implements Extractor {
-                /**
-                 * @param FlowContext $context
-                 *
-                 * @return \Generator<int, Rows, mixed, void>
-                 */
-                public function extract(FlowContext $context) : \Generator
-                {
-                    yield new Rows(
-                        Row::create(new IntegerEntry('id', 1)),
-                        Row::create(new IntegerEntry('id', 2)),
-                        Row::create(new IntegerEntry('id', 3)),
-                        Row::create(new IntegerEntry('id', 4)),
-                        Row::create(new IntegerEntry('id', 5)),
-                        Row::create(new IntegerEntry('id', 6)),
-                        Row::create(new IntegerEntry('id', 7)),
-                        Row::create(new IntegerEntry('id', 8)),
-                        Row::create(new IntegerEntry('id', 9)),
-                        Row::create(new IntegerEntry('id', 10)),
-                    );
-                }
-            }
-        )
-            ->transform(
-                new class implements Transformer {
-                    public function transform(Rows $rows, FlowContext $context) : Rows
-                    {
-                        return $rows->map(fn (Row $row) => $row->rename('id', 'new_id'));
-                    }
-
-                    public function __serialize() : array
-                    {
-                        return [];
-                    }
-
-                    public function __unserialize(array $data) : void
-                    {
-                    }
-                }
-            )
-            ->batchSize(2)
-            ->load(
-                new class implements Loader {
-                    public function load(Rows $rows, FlowContext $context) : void
-                    {
-                        Assert::assertCount(2, $rows);
-                    }
-
-                    public function __serialize() : array
-                    {
-                        return [];
-                    }
-
-                    public function __unserialize(array $data) : void
-                    {
-                    }
-                }
-            )
-            ->run();
-    }
-
-    public function test_etl_with_collecting() : void
-    {
-        (new Flow())->extract(
-            new class implements Extractor {
-                /**
-                 * @param FlowContext $context
-                 *
-                 * @return \Generator<int, Rows, mixed, void>
-                 */
-                public function extract(FlowContext $context) : \Generator
-                {
-                    yield new Rows(Row::create(new IntegerEntry('id', 1)));
-                    yield new Rows(Row::create(new IntegerEntry('id', 2)));
-                    yield new Rows(Row::create(new IntegerEntry('id', 3)));
-                }
-            }
-        )
-            ->transform(
-                new class implements Transformer {
-                    public function transform(Rows $rows, FlowContext $context) : Rows
-                    {
-                        return $rows->map(fn (Row $row) => $row->rename('id', 'new_id'));
-                    }
-
-                    public function __serialize() : array
-                    {
-                        return [];
-                    }
-
-                    public function __unserialize(array $data) : void
-                    {
-                    }
-                }
-            )
-            ->collect()
-            ->load(
-                new class implements Loader {
-                    public function load(Rows $rows, FlowContext $context) : void
-                    {
-                        Assert::assertCount(3, $rows);
-                    }
-
-                    public function __serialize() : array
-                    {
-                        return [];
-                    }
-
-                    public function __unserialize(array $data) : void
-                    {
-                    }
-                }
-            )
-            ->run();
-    }
-
-    public function test_etl_with_total_rows_below_the_limit() : void
-    {
-        $rows = (new Flow())->extract(
-            new class implements Extractor {
-                /**
-                 * @param FlowContext $context
-                 *
-                 * @return \Generator<int, Rows, mixed, void>
-                 */
-                public function extract(FlowContext $context) : \Generator
-                {
-                    for ($i = 0; $i < 5; $i++) {
-                        yield new Rows(
-                            Row::create(new IntegerEntry('id', $i)),
-                        );
-                    }
-                }
-            }
-        )
-            ->limit(10)
-            ->fetch();
-
-        $this->assertCount(5, $rows);
-    }
-
-    public function test_fetch_with_limit_below_0() : void
-    {
-        $this->expectException(InvalidArgumentException::class);
-        $this->expectExceptionMessage("Limit can't be lower or equal zero, given: -1");
-
-        (new Flow())->process(new Rows())
-            ->fetch(-1);
     }
 
     public function test_filter_partitions() : void
@@ -1048,256 +429,30 @@ ASCIITABLE,
         );
     }
 
-    public function test_group_by_multiple_columns_and_batch_size() : void
+    public function test_limit() : void
     {
-        $loader = $this->createMock(Loader::class);
-        $loader->expects($this->exactly(4))
-            ->method('load');
-
-        $rows = (new Flow())->process(
-            new Rows(
-                Row::create(Entry::integer('id', 1), Entry::string('country', 'PL'), Entry::integer('age', 20), Entry::string('gender', 'male')),
-                Row::create(Entry::integer('id', 2), Entry::string('country', 'PL'), Entry::integer('age', 20), Entry::string('gender', 'male')),
-                Row::create(Entry::integer('id', 3), Entry::string('country', 'PL'), Entry::integer('age', 25), Entry::string('gender', 'male')),
-                Row::create(Entry::integer('id', 4), Entry::string('country', 'PL'), Entry::integer('age', 30), Entry::string('gender', 'female')),
-                Row::create(Entry::integer('id', 5), Entry::string('country', 'US'), Entry::integer('age', 40), Entry::string('gender', 'female')),
-                Row::create(Entry::integer('id', 6), Entry::string('country', 'US'), Entry::integer('age', 40), Entry::string('gender', 'male')),
-                Row::create(Entry::integer('id', 7), Entry::string('country', 'US'), Entry::integer('age', 45), Entry::string('gender', 'female')),
-                Row::create(Entry::integer('id', 9), Entry::string('country', 'US'), Entry::integer('age', 50), Entry::string('gender', 'male')),
-            )
-        )
-            ->groupBy('country', 'gender')
-            ->batchSize(1)
-            ->write($loader)
-            ->fetch();
-
-        $this->assertEquals(
-            new Rows(
-                Row::create(Entry::string('country', 'PL'), Entry::string('gender', 'male')),
-                Row::create(Entry::string('country', 'PL'), Entry::string('gender', 'female')),
-                Row::create(Entry::string('country', 'US'), Entry::string('gender', 'female')),
-                Row::create(Entry::string('country', 'US'), Entry::string('gender', 'male')),
-            ),
-            $rows
-        );
-    }
-
-    public function test_group_by_multiples_columns_with_avg_aggregation() : void
-    {
-        $rows = (new Flow())->process(
-            new Rows(
-                Row::create(Entry::integer('id', 1), Entry::string('country', 'PL'), Entry::integer('age', 20), Entry::string('gender', 'male')),
-                Row::create(Entry::integer('id', 2), Entry::string('country', 'PL'), Entry::integer('age', 20), Entry::string('gender', 'male')),
-                Row::create(Entry::integer('id', 3), Entry::string('country', 'PL'), Entry::integer('age', 25), Entry::string('gender', 'male')),
-                Row::create(Entry::integer('id', 4), Entry::string('country', 'PL'), Entry::integer('age', 30), Entry::string('gender', 'female')),
-                Row::create(Entry::integer('id', 5), Entry::string('country', 'US'), Entry::integer('age', 40), Entry::string('gender', 'female')),
-                Row::create(Entry::integer('id', 6), Entry::string('country', 'US'), Entry::integer('age', 40), Entry::string('gender', 'male')),
-                Row::create(Entry::integer('id', 7), Entry::string('country', 'US'), Entry::integer('age', 45), Entry::string('gender', 'female')),
-                Row::create(Entry::integer('id', 9), Entry::string('country', 'US'), Entry::integer('age', 50), Entry::string('gender', 'male')),
-            )
-        )
-            ->groupBy('country', 'gender')
-            ->aggregate(average(ref('age')))
-            ->fetch();
-
-        $this->assertEquals(
-            new Rows(
-                Row::create(Entry::string('country', 'PL'), Entry::string('gender', 'male'), Entry::float('age_avg', 21.666666666666668)),
-                Row::create(Entry::string('country', 'PL'), Entry::string('gender', 'female'), Entry::integer('age_avg', 30)),
-                Row::create(Entry::string('country', 'US'), Entry::string('gender', 'female'), Entry::float('age_avg', 42.5)),
-                Row::create(Entry::string('country', 'US'), Entry::string('gender', 'male'), Entry::integer('age_avg', 45)),
-            ),
-            $rows
-        );
-    }
-
-    public function test_group_by_multiples_columns_with_avg_aggregation_with_null() : void
-    {
-        $rows = (new Flow())->process(
-            new Rows(
-                Row::create(Entry::integer('id', 1), Entry::string('country', 'PL'), Entry::integer('age', 20), Entry::string('gender', 'male')),
-                Row::create(Entry::integer('id', 2), Entry::string('country', 'PL'), Entry::integer('age', 20), Entry::string('gender', 'male')),
-                Row::create(Entry::integer('id', 3), Entry::string('country', 'PL'), Entry::integer('age', 25), Entry::string('gender', 'male')),
-                Row::create(Entry::integer('id', 4), Entry::string('country', 'PL'), Entry::integer('age', 30), Entry::string('gender', 'female')),
-                Row::create(Entry::integer('id', 5), Entry::string('country', 'US'), Entry::integer('age', 40), Entry::string('gender', 'female')),
-                Row::create(Entry::integer('id', 6), Entry::string('country', 'US'), Entry::integer('age', 40), Entry::string('gender', 'male')),
-                Row::create(Entry::integer('id', 7), Entry::string('country', 'US'), Entry::integer('age', 45), Entry::null('gender')),
-                Row::create(Entry::integer('id', 9), Entry::string('country', 'US'), Entry::integer('age', 50), Entry::string('gender', 'male')),
-            )
-        )
-            ->groupBy('country', 'gender')
-            ->aggregate(average(ref('age')))
-            ->fetch();
-
-        $this->assertEquals(
-            new Rows(
-                Row::create(Entry::string('country', 'PL'), Entry::string('gender', 'male'), Entry::float('age_avg', 21.666666666666668)),
-                Row::create(Entry::string('country', 'PL'), Entry::string('gender', 'female'), Entry::integer('age_avg', 30)),
-                Row::create(Entry::string('country', 'US'), Entry::string('gender', 'female'), Entry::integer('age_avg', 40)),
-                Row::create(Entry::string('country', 'US'), Entry::string('gender', 'male'), Entry::integer('age_avg', 45)),
-                Row::create(Entry::string('country', 'US'), Entry::null('gender'), Entry::integer('age_avg', 45)),
-            ),
-            $rows
-        );
-    }
-
-    public function test_group_by_single_column() : void
-    {
-        $rows = (new Flow())->process(
-            new Rows(
-                Row::create(Entry::integer('id', 1), Entry::string('country', 'PL'), Entry::integer('age', 20)),
-                Row::create(Entry::integer('id', 2), Entry::string('country', 'PL'), Entry::integer('age', 20)),
-                Row::create(Entry::integer('id', 3), Entry::string('country', 'PL'), Entry::integer('age', 25)),
-                Row::create(Entry::integer('id', 4), Entry::string('country', 'PL'), Entry::integer('age', 30)),
-                Row::create(Entry::integer('id', 5), Entry::string('country', 'US'), Entry::integer('age', 40)),
-                Row::create(Entry::integer('id', 6), Entry::string('country', 'US'), Entry::integer('age', 40)),
-                Row::create(Entry::integer('id', 7), Entry::string('country', 'US'), Entry::integer('age', 45)),
-                Row::create(Entry::integer('id', 9), Entry::string('country', 'US'), Entry::integer('age', 50)),
-            )
-        )
-            ->groupBy('country')
-            ->fetch();
-
-        $this->assertEquals(
-            new Rows(
-                Row::create(Entry::string('country', 'PL')),
-                Row::create(Entry::string('country', 'US')),
-            ),
-            $rows
-        );
-    }
-
-    public function test_group_by_single_column_with_avg_aggregation() : void
-    {
-        $rows = (new Flow())->process(
-            new Rows(
-                Row::create(Entry::integer('id', 1), Entry::string('country', 'PL'), Entry::integer('age', 20)),
-                Row::create(Entry::integer('id', 2), Entry::string('country', 'PL'), Entry::integer('age', 20)),
-                Row::create(Entry::integer('id', 3), Entry::string('country', 'PL'), Entry::integer('age', 25)),
-                Row::create(Entry::integer('id', 4), Entry::string('country', 'PL'), Entry::integer('age', 30)),
-                Row::create(Entry::integer('id', 5), Entry::string('country', 'US'), Entry::integer('age', 40)),
-                Row::create(Entry::integer('id', 6), Entry::string('country', 'US'), Entry::integer('age', 40)),
-                Row::create(Entry::integer('id', 7), Entry::string('country', 'US'), Entry::integer('age', 45)),
-                Row::create(Entry::integer('id', 9), Entry::string('country', 'US'), Entry::integer('age', 50)),
-            )
-        )
-            ->groupBy('country')
-            ->aggregate(average(ref('age')))
-            ->fetch();
-
-        $this->assertEquals(
-            new Rows(
-                Row::create(Entry::string('country', 'PL'), Entry::float('age_avg', 23.75)),
-                Row::create(Entry::string('country', 'US'), Entry::float('age_avg', 43.75)),
-            ),
-            $rows
-        );
-    }
-
-    public function test_join() : void
-    {
-        $loader = $this->createMock(Loader::class);
-        $loader->expects($this->exactly(2))
-            ->method('load');
-
-        $rows = (new Flow())->process(
-            new Rows(
-                Row::create(Entry::integer('id', 1), Entry::string('country', 'PL')),
-                Row::create(Entry::integer('id', 2), Entry::string('country', 'PL')),
-                Row::create(Entry::integer('id', 3), Entry::string('country', 'PL')),
-                Row::create(Entry::integer('id', 4), Entry::string('country', 'PL')),
-                Row::create(Entry::integer('id', 5), Entry::string('country', 'US')),
-                Row::create(Entry::integer('id', 6), Entry::string('country', 'US')),
-                Row::create(Entry::integer('id', 7), Entry::string('country', 'US')),
-                Row::create(Entry::integer('id', 9), Entry::string('country', 'US')),
-            )
-        )
-            ->batchSize(4)
-            ->join(
-                (new Flow())->process(
-                    new Rows(
-                        Row::create(Entry::string('code', 'PL'), Entry::string('name', 'Poland')),
-                        Row::create(Entry::string('code', 'US'), Entry::string('name', 'United States')),
-                    )
-                ),
-                Expression::on(['country' => 'code']),
-            )
-            ->write($loader)
-            ->fetch();
-
-        $this->assertEquals(
-            new Rows(
-                Row::create(Entry::integer('id', 1), Entry::string('country', 'PL'), Entry::string('name', 'Poland')),
-                Row::create(Entry::integer('id', 2), Entry::string('country', 'PL'), Entry::string('name', 'Poland')),
-                Row::create(Entry::integer('id', 3), Entry::string('country', 'PL'), Entry::string('name', 'Poland')),
-                Row::create(Entry::integer('id', 4), Entry::string('country', 'PL'), Entry::string('name', 'Poland')),
-                Row::create(Entry::integer('id', 5), Entry::string('country', 'US'), Entry::string('name', 'United States')),
-                Row::create(Entry::integer('id', 6), Entry::string('country', 'US'), Entry::string('name', 'United States')),
-                Row::create(Entry::integer('id', 7), Entry::string('country', 'US'), Entry::string('name', 'United States')),
-                Row::create(Entry::integer('id', 9), Entry::string('country', 'US'), Entry::string('name', 'United States')),
-            ),
-            $rows
-        );
-    }
-
-    public function test_join_each() : void
-    {
-        $loader = $this->createMock(Loader::class);
-        $loader->expects($this->exactly(2))
-            ->method('load');
-
-        $rows = (new Flow())->process(
-            new Rows(
-                Row::create(Entry::integer('id', 1), Entry::string('country', 'PL')),
-                Row::create(Entry::integer('id', 2), Entry::string('country', 'PL')),
-                Row::create(Entry::integer('id', 3), Entry::string('country', 'PL')),
-                Row::create(Entry::integer('id', 4), Entry::string('country', 'PL')),
-                Row::create(Entry::integer('id', 5), Entry::string('country', 'US')),
-                Row::create(Entry::integer('id', 6), Entry::string('country', 'US')),
-                Row::create(Entry::integer('id', 7), Entry::string('country', 'US')),
-                Row::create(Entry::integer('id', 9), Entry::string('country', 'US')),
-            )
-        )
-            ->batchSize(4)
-            ->joinEach(
-                new class implements DataFrameFactory {
-                    public function from(Rows $rows) : DataFrame
-                    {
-                        return (new Flow())->process(
-                            new Rows(
-                                Row::create(Entry::string('code', 'PL'), Entry::string('name', 'Poland')),
-                                Row::create(Entry::string('code', 'US'), Entry::string('name', 'United States')),
-                            )
+        $rows = (new Flow())->extract(
+            new class implements Extractor {
+                /**
+                 * @param FlowContext $context
+                 *
+                 * @return \Generator<int, Rows, mixed, void>
+                 */
+                public function extract(FlowContext $context) : \Generator
+                {
+                    for ($i = 0; $i < 1000; $i++) {
+                        yield new Rows(
+                            Row::create(new IntegerEntry('id', $i + 1)),
+                            Row::create(new IntegerEntry('id', $i + 2)),
                         );
                     }
+                }
+            }
+        )
+        ->limit(10)
+        ->fetch();
 
-                    public function __serialize() : array
-                    {
-                        return [];
-                    }
-
-                    public function __unserialize(array $data) : void
-                    {
-                    }
-                },
-                Expression::on(['country' => 'code']),
-            )
-            ->write($loader)
-            ->fetch();
-
-        $this->assertEquals(
-            new Rows(
-                Row::create(Entry::integer('id', 1), Entry::string('country', 'PL'), Entry::string('name', 'Poland')),
-                Row::create(Entry::integer('id', 2), Entry::string('country', 'PL'), Entry::string('name', 'Poland')),
-                Row::create(Entry::integer('id', 3), Entry::string('country', 'PL'), Entry::string('name', 'Poland')),
-                Row::create(Entry::integer('id', 4), Entry::string('country', 'PL'), Entry::string('name', 'Poland')),
-                Row::create(Entry::integer('id', 5), Entry::string('country', 'US'), Entry::string('name', 'United States')),
-                Row::create(Entry::integer('id', 6), Entry::string('country', 'US'), Entry::string('name', 'United States')),
-                Row::create(Entry::integer('id', 7), Entry::string('country', 'US'), Entry::string('name', 'United States')),
-                Row::create(Entry::integer('id', 9), Entry::string('country', 'US'), Entry::string('name', 'United States')),
-            ),
-            $rows
-        );
+        $this->assertCount(10, $rows);
     }
 
     public function test_limit_below_0() : void
@@ -1345,30 +500,97 @@ ASCIITABLE,
         $this->assertCount(3, $rows);
     }
 
-    public function test_overriding_group_by() : void
+    public function test_limit_with_batch_size() : void
     {
-        $rows = (new Flow())->process(
-            new Rows(
-                Row::create(Entry::integer('id', 1), Entry::string('country', 'PL'), Entry::integer('age', 20)),
-                Row::create(Entry::integer('id', 2), Entry::string('country', 'PL'), Entry::integer('age', 20)),
-                Row::create(Entry::integer('id', 3), Entry::string('country', 'PL'), Entry::integer('age', 25)),
-                Row::create(Entry::integer('id', 4), Entry::string('country', 'PL'), Entry::integer('age', 30)),
-                Row::create(Entry::integer('id', 5), Entry::string('country', 'US'), Entry::integer('age', 40)),
-                Row::create(Entry::integer('id', 6), Entry::string('country', 'US'), Entry::integer('age', 40)),
-                Row::create(Entry::integer('id', 7), Entry::string('country', 'US'), Entry::integer('age', 45)),
-                Row::create(Entry::integer('id', 9), Entry::string('country', 'US'), Entry::integer('age', 50)),
-            )
+        $rows = (new Flow())->extract(
+            new class implements Extractor {
+                /**
+                 * @param FlowContext $context
+                 *
+                 * @return \Generator<int, Rows, mixed, void>
+                 */
+                public function extract(FlowContext $context) : \Generator
+                {
+                    for ($i = 0; $i < 1000; $i++) {
+                        yield new Rows(
+                            Row::create(new IntegerEntry('id', $i + 1)),
+                            Row::create(new IntegerEntry('id', $i + 2)),
+                        );
+                    }
+                }
+            }
         )
-            ->groupBy('id')
-            ->groupBy('country')
+            ->batchSize(50)
+            ->limit(10)
             ->fetch();
 
-        $this->assertEquals(
-            new Rows(
-                Row::create(Entry::string('country', 'PL')),
-                Row::create(Entry::string('country', 'US')),
-            ),
-            $rows
+        $this->assertCount(10, $rows);
+    }
+
+    public function test_limit_with_collecting() : void
+    {
+        $rows = (new Flow())->extract(
+            new class implements Extractor {
+                /**
+                 * @param FlowContext $context
+                 *
+                 * @return \Generator<int, Rows, mixed, void>
+                 */
+                public function extract(FlowContext $context) : \Generator
+                {
+                    for ($i = 0; $i < 1000; $i++) {
+                        yield new Rows(
+                            Row::create(new IntegerEntry('id', $i + 1)),
+                            Row::create(new IntegerEntry('id', $i + 2)),
+                        );
+                    }
+                }
+            }
+        )
+            ->limit(10)
+            ->collect()
+            ->fetch();
+
+        $this->assertCount(10, $rows);
+    }
+
+    public function test_map() : void
+    {
+        $rows = (new Flow())->extract(
+            new class implements Extractor {
+                /**
+                 * @param FlowContext $context
+                 *
+                 * @return \Generator<int, Rows, mixed, void>
+                 */
+                public function extract(FlowContext $context) : \Generator
+                {
+                    for ($i = 1; $i <= 10; $i++) {
+                        yield new Rows(
+                            Row::create(new IntegerEntry('id', $i)),
+                        );
+                    }
+                }
+            }
+        )
+            ->map(fn (Row $row) => $row->add(new BooleanEntry('odd', $row->valueOf('id') % 2 === 0)))
+            ->fetch();
+
+        $this->assertCount(10, $rows);
+        $this->assertSame(
+            [
+                ['id' => 1, 'odd' => false],
+                ['id' => 2, 'odd' => true],
+                ['id' => 3, 'odd' => false],
+                ['id' => 4, 'odd' => true],
+                ['id' => 5, 'odd' => false],
+                ['id' => 6, 'odd' => true],
+                ['id' => 7, 'odd' => false],
+                ['id' => 8, 'odd' => true],
+                ['id' => 9, 'odd' => false],
+                ['id' => 10, 'odd' => true],
+            ],
+            $rows->toArray()
         );
     }
 
@@ -1413,136 +635,132 @@ ASCIITABLE,
         );
     }
 
-    public function test_pivot() : void
+    public function test_pipeline() : void
     {
-        $dataset1 = [
-            ['date' => '2023-11-01', 'user' => 'norberttech', 'contributions' => 5],
-            ['date' => '2023-11-01', 'user' => 'stloyd', 'contributions' => 4],
-            ['date' => '2023-11-02', 'user' => 'norberttech', 'contributions' => 3],
-            ['date' => '2023-11-02', 'user' => 'stloyd', 'contributions' => 6],
-        ];
+        $extractor = new class implements Extractor {
+            /**
+             * @param FlowContext $context
+             *
+             * @return \Generator<int, Rows, mixed, void>
+             */
+            public function extract(FlowContext $context) : \Generator
+            {
+                yield new Rows(
+                    Row::create(
+                        new IntegerEntry('id', 101),
+                        new BooleanEntry('deleted', false),
+                        new DateTimeEntry('expiration-date', new \DateTimeImmutable('2020-08-24')),
+                        new NullEntry('phase')
+                    )
+                );
 
-        $dataset2 = [
-            ['date' => '2023-11-03', 'user' => 'norberttech', 'contributions' => 2],
-            ['date' => '2023-11-03', 'user' => 'stloyd', 'contributions' => 7],
-            ['date' => '2023-11-04', 'user' => 'norberttech', 'contributions' => 3],
-            ['date' => '2023-11-04', 'user' => 'stloyd', 'contributions' => 5],
-            ['date' => '2023-11-05', 'user' => 'norberttech', 'contributions' => 7],
-            ['date' => '2023-11-05', 'user' => 'stloyd', 'contributions' => 11],
-        ];
+                yield new Rows(
+                    Row::create(
+                        new IntegerEntry('id', 102),
+                        new BooleanEntry('deleted', true),
+                        new DateTimeEntry('expiration-date', new \DateTimeImmutable('2020-08-25')),
+                        new NullEntry('phase')
+                    )
+                );
+            }
+        };
 
-        $rows = (new Flow())
-            ->read(
-                From::chain(
-                    From::array($dataset1),
-                    From::array($dataset2),
-                )
-            )
-            ->groupBy(ref('date'))
-            ->pivot(ref('user'))
-            ->aggregate(sum(ref('contributions')))
-            ->fetch();
+        $addStampStringEntry = new class implements Transformer {
+            public function transform(Rows $rows, FlowContext $context) : Rows
+            {
+                return $rows->map(
+                    fn (Row $row) : Row => $row->set(new StringEntry('stamp', 'zero'))
+                );
+            }
 
-        $this->assertSame(
+            public function __serialize() : array
+            {
+                return [];
+            }
+
+            public function __unserialize(array $data) : void
+            {
+            }
+        };
+
+        $loader = new class implements Loader {
+            public array $result = [];
+
+            public function load(Rows $rows, FlowContext $context) : void
+            {
+                $this->result = \array_merge($this->result, $rows->toArray());
+            }
+
+            public function __serialize() : array
+            {
+                return [];
+            }
+
+            public function __unserialize(array $data) : void
+            {
+            }
+        };
+
+        (new Flow())->read($extractor)
+            ->onError(new IgnoreError())
+            ->rows($addStampStringEntry)
+            ->rows(new class implements Transformer {
+                public function transform(Rows $rows, FlowContext $context) : Rows
+                {
+                    throw new \RuntimeException('Unexpected exception');
+                }
+
+                public function __serialize() : array
+                {
+                    return [];
+                }
+
+                public function __unserialize(array $data) : void
+                {
+                }
+            })
+            ->rows(AddStampToStringEntryTransformer::divideBySemicolon('stamp', 'one'))
+            ->rows(AddStampToStringEntryTransformer::divideBySemicolon('stamp', 'two'))
+            ->rows(AddStampToStringEntryTransformer::divideBySemicolon('stamp', 'three'))
+            ->write($loader)
+            ->run();
+
+        $this->assertEquals(
             [
                 [
-                    'date' => '2023-11-01',
-                    'norberttech' => 5,
-                    'stloyd' => 4,
+                    'id' => 101,
+                    'stamp' => 'zero:one:two:three',
+                    'deleted' => false,
+                    'expiration-date' => new \DateTimeImmutable('2020-08-24'),
+                    'phase' => null,
                 ],
                 [
-                    'date' => '2023-11-02',
-                    'norberttech' => 3,
-                    'stloyd' => 6,
-                ],
-                [
-                    'date' => '2023-11-03',
-                    'norberttech' => 2,
-                    'stloyd' => 7,
-                ],
-                [
-                    'date' => '2023-11-04',
-                    'norberttech' => 3,
-                    'stloyd' => 5,
-                ],
-                [
-                    'date' => '2023-11-05',
-                    'norberttech' => 7,
-                    'stloyd' => 11,
+                    'id' => 102,
+                    'stamp' => 'zero:one:two:three',
+                    'deleted' => true,
+                    'expiration-date' => new \DateTimeImmutable('2020-08-25'),
+                    'phase' => null,
                 ],
             ],
-            $rows->toArray()
+            $loader->result,
         );
     }
 
-    public function test_print_rows() : void
+    public function test_process_constructor() : void
     {
-        \ob_start();
-        (new Flow())->process(
-            new Rows(
-                Row::create(Entry::integer('id', 1), Entry::string('country', 'PL'), Entry::integer('age', 20)),
-                Row::create(Entry::integer('id', 2), Entry::string('country', 'PL'), Entry::integer('age', 20)),
-                Row::create(Entry::integer('id', 3), Entry::string('country', 'PL'), Entry::integer('age', 25)),
-            ),
-            new Rows(
-                Row::create(Entry::integer('id', 1), Entry::string('country', 'PL'), Entry::integer('age', 20), Entry::integer('salary', 5000)),
-                Row::create(Entry::integer('id', 1), Entry::string('country', 'PL'), Entry::integer('age', 20), Entry::null('salary')),
+        $collectedRows = (new Flow())->process(
+            $rows = new Rows(
+                Row::create(
+                    new IntegerEntry('id', 101),
+                    new BooleanEntry('deleted', false),
+                    new DateTimeEntry('expiration-date', new \DateTimeImmutable('2020-08-24')),
+                    new NullEntry('phase')
+                )
             )
-        )->printRows();
-        $output = \ob_get_clean();
+        )
+            ->fetch();
 
-        $this->assertStringContainsString(
-            <<<'ASCII'
-+----+---------+-----+
-| id | country | age |
-+----+---------+-----+
-|  1 |      PL |  20 |
-|  2 |      PL |  20 |
-|  3 |      PL |  25 |
-+----+---------+-----+
-3 rows
-+----+---------+-----+--------+
-| id | country | age | salary |
-+----+---------+-----+--------+
-|  1 |      PL |  20 |   5000 |
-|  1 |      PL |  20 |   null |
-+----+---------+-----+--------+
-2 rows
-ASCII,
-            $output
-        );
-    }
-
-    public function test_print_schema() : void
-    {
-        \ob_start();
-        (new Flow())->process(
-            new Rows(
-                Row::create(Entry::integer('id', 1), Entry::string('country', 'PL'), Entry::integer('age', 20)),
-                Row::create(Entry::integer('id', 2), Entry::string('country', 'PL'), Entry::integer('age', 20)),
-                Row::create(Entry::integer('id', 3), Entry::string('country', 'PL'), Entry::integer('age', 25)),
-            ),
-            new Rows(
-                Row::create(Entry::integer('id', 1), Entry::string('country', 'PL'), Entry::integer('age', 20), Entry::integer('salary', 5000)),
-                Row::create(Entry::integer('id', 1), Entry::string('country', 'PL'), Entry::integer('age', 20), Entry::null('salary')),
-            )
-        )->printSchema();
-        $output = \ob_get_clean();
-
-        $this->assertStringContainsString(
-            <<<'ASCII'
-schema
-|-- id: integer
-|-- country: string
-|-- age: integer
-schema
-|-- id: integer
-|-- country: string
-|-- age: integer
-|-- salary: ?integer
-ASCII,
-            $output
-        );
+        $this->assertEquals($rows, $collectedRows);
     }
 
     public function test_rename() : void
@@ -1741,60 +959,6 @@ ASCII,
         );
     }
 
-    public function test_standalone_avg_aggregation() : void
-    {
-        $rows = (new Flow())->process(
-            new Rows(
-                Row::create(Entry::integer('id', 1), Entry::string('country', 'PL'), Entry::integer('age', 20)),
-                Row::create(Entry::integer('id', 2), Entry::string('country', 'PL'), Entry::integer('age', 20)),
-                Row::create(Entry::integer('id', 3), Entry::string('country', 'PL'), Entry::integer('age', 25)),
-                Row::create(Entry::integer('id', 4), Entry::string('country', 'PL'), Entry::integer('age', 30)),
-                Row::create(Entry::integer('id', 5), Entry::string('country', 'US'), Entry::integer('age', 40)),
-                Row::create(Entry::integer('id', 6), Entry::string('country', 'US'), Entry::integer('age', 40)),
-                Row::create(Entry::integer('id', 7), Entry::string('country', 'US'), Entry::integer('age', 45)),
-                Row::create(Entry::integer('id', 9), Entry::string('country', 'US'), Entry::integer('age', 50)),
-            )
-        )
-            ->aggregate(average(ref('age')))
-            ->rows(Transform::rename('age_avg', 'average_age'))
-            ->fetch();
-
-        $this->assertEquals(
-            new Rows(
-                Row::create(Entry::float('average_age', 33.75)),
-            ),
-            $rows
-        );
-    }
-
-    public function test_standalone_avg_and_max_aggregation() : void
-    {
-        (new Flow())->process(
-            new Rows(
-                Row::create(Entry::integer('id', 1), Entry::string('country', 'PL'), Entry::integer('age', 20)),
-                Row::create(Entry::integer('id', 2), Entry::string('country', 'PL'), Entry::integer('age', 20)),
-                Row::create(Entry::integer('id', 3), Entry::string('country', 'PL'), Entry::integer('age', 25)),
-                Row::create(Entry::integer('id', 4), Entry::string('country', 'PL'), Entry::integer('age', 30)),
-                Row::create(Entry::integer('id', 5), Entry::string('country', 'US'), Entry::integer('age', 40)),
-                Row::create(Entry::integer('id', 6), Entry::string('country', 'US'), Entry::integer('age', 40)),
-                Row::create(Entry::integer('id', 7), Entry::string('country', 'US'), Entry::integer('age', 45)),
-                Row::create(Entry::integer('id', 9), Entry::string('country', 'US'), Entry::integer('age', 50)),
-            )
-        )
-            ->aggregate(average(ref('age')), max(ref('age')))
-            ->run(function (Rows $rows) : void {
-                $this->assertEquals(
-                    new Rows(
-                        Row::create(
-                            Entry::float('age_avg', 33.75),
-                            Entry::integer('age_max', 50)
-                        )
-                    ),
-                    $rows
-                );
-            });
-    }
-
     public function test_strict_validation_against_schema() : void
     {
         $rows = (new Flow())->process(
@@ -1880,85 +1044,147 @@ ASCII,
         );
     }
 
-    public function test_window_avg_function() : void
+    public function test_with_batch_size() : void
     {
-        $memoryPage1 = new ArrayMemory([
-            ['employee_name' => 'James', 'department' => 'Sales', 'salary' => 3000],
-            ['employee_name' => 'Michael', 'department' => 'Sales', 'salary' => 4600],
-            ['employee_name' => 'Jeff', 'department' => 'Marketing', 'salary' => 3000],
-            ['employee_name' => 'Saif', 'department' => 'Sales', 'salary' => 4100],
-            ['employee_name' => 'John', 'department' => 'Marketing', 'salary' => 3200],
-        ]);
-        $memoryPage2 = new ArrayMemory([
-            ['employee_name' => 'Emma', 'department' => 'Sales', 'salary' => 4800],
-            ['employee_name' => 'Oliver', 'department' => 'Sales', 'salary' => 2900],
-            ['employee_name' => 'Mia', 'department' => 'Finance', 'salary' => 3300],
-            ['employee_name' => 'Noah', 'department' => 'Marketing', 'salary' => 3400],
-            ['employee_name' => 'Ava', 'department' => 'Finance', 'salary' => 3800],
-            ['employee_name' => 'Isabella', 'department' => 'Marketing', 'salary' => 2100],
-            ['employee_name' => 'Ethan', 'department' => 'Sales', 'salary' => 4100],
-            ['employee_name' => 'Charlotte', 'department' => 'Marketing', 'salary' => 3000],
-        ]);
+        (new Flow())->extract(
+            new class implements Extractor {
+                /**
+                 * @param FlowContext $context
+                 *
+                 * @return \Generator<int, Rows, mixed, void>
+                 */
+                public function extract(FlowContext $context) : \Generator
+                {
+                    yield new Rows(
+                        Row::create(new IntegerEntry('id', 1)),
+                        Row::create(new IntegerEntry('id', 2)),
+                        Row::create(new IntegerEntry('id', 3)),
+                        Row::create(new IntegerEntry('id', 4)),
+                        Row::create(new IntegerEntry('id', 5)),
+                        Row::create(new IntegerEntry('id', 6)),
+                        Row::create(new IntegerEntry('id', 7)),
+                        Row::create(new IntegerEntry('id', 8)),
+                        Row::create(new IntegerEntry('id', 9)),
+                        Row::create(new IntegerEntry('id', 10)),
+                    );
+                }
+            }
+        )
+            ->transform(
+                new class implements Transformer {
+                    public function transform(Rows $rows, FlowContext $context) : Rows
+                    {
+                        return $rows->map(fn (Row $row) => $row->rename('id', 'new_id'));
+                    }
 
-        $this->assertSame(
-            [
-                ['department' => 'Sales', 'avg_salary' => 3917.0],
-                ['department' => 'Marketing', 'avg_salary' => 2940.0],
-                ['department' => 'Finance', 'avg_salary' => 3550.0],
-            ],
-            (new Flow)
-                ->read(From::chain(From::memory($memoryPage1), From::memory($memoryPage2)))
-                ->withEntry('avg_salary', average(ref('salary'))->over(window()->partitionBy(ref('department'))))
-                ->select('department', 'avg_salary')
-                ->dropDuplicates(ref('department'), ref('avg_salary'))
-                ->withEntry('avg_salary', ref('avg_salary')->round(lit(0)))
-                ->fetch()
-                ->toArray()
-        );
+                    public function __serialize() : array
+                    {
+                        return [];
+                    }
+
+                    public function __unserialize(array $data) : void
+                    {
+                    }
+                }
+            )
+            ->batchSize(2)
+            ->load(
+                new class implements Loader {
+                    public function load(Rows $rows, FlowContext $context) : void
+                    {
+                        Assert::assertCount(2, $rows);
+                    }
+
+                    public function __serialize() : array
+                    {
+                        return [];
+                    }
+
+                    public function __unserialize(array $data) : void
+                    {
+                    }
+                }
+            )
+            ->run();
     }
 
-    public function test_window_rank_function() : void
+    public function test_with_collecting() : void
     {
-        $memoryPage1 = new ArrayMemory([
-            ['employee_name' => 'James', 'department' => 'Sales', 'salary' => 3000],
-            ['employee_name' => 'James', 'department' => 'Sales', 'salary' => 3000],
-            ['employee_name' => 'James', 'department' => 'Sales', 'salary' => 3000],
-            ['employee_name' => 'Michael', 'department' => 'Sales', 'salary' => 4600],
-            ['employee_name' => 'Robert', 'department' => 'Sales', 'salary' => 4100],
-            ['employee_name' => 'Maria', 'department' => 'Finance', 'salary' => 3000],
-            ['employee_name' => 'Scott', 'department' => 'Finance', 'salary' => 3300],
-            ['employee_name' => 'Jen', 'department' => 'Finance', 'salary' => 3900],
-            ['employee_name' => 'Jeff', 'department' => 'Marketing', 'salary' => 3000],
-            ['employee_name' => 'Kumar', 'department' => 'Marketing', 'salary' => 2000],
-            ['employee_name' => 'Saif', 'department' => 'Sales', 'salary' => 4100],
-            ['employee_name' => 'John', 'department' => 'Marketing', 'salary' => 3200],
-        ]);
-        $memoryPage2 = new ArrayMemory([
-            ['employee_name' => 'Emma', 'department' => 'Sales', 'salary' => 4800],
-            ['employee_name' => 'Sophia', 'department' => 'Finance', 'salary' => 4200],
-            ['employee_name' => 'Oliver', 'department' => 'Sales', 'salary' => 2900],
-            ['employee_name' => 'Mia', 'department' => 'Finance', 'salary' => 3300],
-            ['employee_name' => 'Noah', 'department' => 'Marketing', 'salary' => 3400],
-            ['employee_name' => 'Ava', 'department' => 'Finance', 'salary' => 3800],
-            ['employee_name' => 'Liam', 'department' => 'Sales', 'salary' => 3100],
-            ['employee_name' => 'Isabella', 'department' => 'Marketing', 'salary' => 2100],
-            ['employee_name' => 'Ethan', 'department' => 'Sales', 'salary' => 4100],
-            ['employee_name' => 'Charlotte', 'department' => 'Marketing', 'salary' => 3000],
-        ]);
+        (new Flow())->extract(
+            new class implements Extractor {
+                /**
+                 * @param FlowContext $context
+                 *
+                 * @return \Generator<int, Rows, mixed, void>
+                 */
+                public function extract(FlowContext $context) : \Generator
+                {
+                    yield new Rows(Row::create(new IntegerEntry('id', 1)));
+                    yield new Rows(Row::create(new IntegerEntry('id', 2)));
+                    yield new Rows(Row::create(new IntegerEntry('id', 3)));
+                }
+            }
+        )
+            ->transform(
+                new class implements Transformer {
+                    public function transform(Rows $rows, FlowContext $context) : Rows
+                    {
+                        return $rows->map(fn (Row $row) => $row->rename('id', 'new_id'));
+                    }
 
-        $this->assertSame(
-            [
-                ['employee_name' => 'Emma', 'department' => 'Sales', 'salary' => 4800, 'rank' => 1],
-                ['employee_name' => 'Sophia', 'department' => 'Finance', 'salary' => 4200, 'rank' => 1],
-                ['employee_name' => 'Noah', 'department' => 'Marketing', 'salary' => 3400, 'rank' => 1],
-            ],
-            (new Flow)
-                ->read(From::all(From::memory($memoryPage1), From::memory($memoryPage2)))
-                ->dropDuplicates(ref('employee_name'), ref('department'))
-                ->withEntry('rank', rank()->over(window()->partitionBy(ref('department'))->orderBy(ref('salary')->desc())))
-                ->filter(ref('rank')->equals(lit(1)))
-                ->fetch()
-                ->toArray()
-        );
+                    public function __serialize() : array
+                    {
+                        return [];
+                    }
+
+                    public function __unserialize(array $data) : void
+                    {
+                    }
+                }
+            )
+            ->collect()
+            ->load(
+                new class implements Loader {
+                    public function load(Rows $rows, FlowContext $context) : void
+                    {
+                        Assert::assertCount(3, $rows);
+                    }
+
+                    public function __serialize() : array
+                    {
+                        return [];
+                    }
+
+                    public function __unserialize(array $data) : void
+                    {
+                    }
+                }
+            )
+            ->run();
+    }
+
+    public function test_with_total_rows_below_the_limit() : void
+    {
+        $rows = (new Flow())->extract(
+            new class implements Extractor {
+                /**
+                 * @param FlowContext $context
+                 *
+                 * @return \Generator<int, Rows, mixed, void>
+                 */
+                public function extract(FlowContext $context) : \Generator
+                {
+                    for ($i = 0; $i < 5; $i++) {
+                        yield new Rows(
+                            Row::create(new IntegerEntry('id', $i)),
+                        );
+                    }
+                }
+            }
+        )
+            ->limit(10)
+            ->fetch();
+
+        $this->assertCount(5, $rows);
     }
 }
